api/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.2.1
langchain-groq==1.0.0
langchain-core==1.0.3
langchain-community
langchain-text-splitters
langchain
requests
huggingface-hub
faiss-cpu
